UruguayLands/requirements.txt
# Основные зависимости
pydantic>=2.0.0
playwright>=1.40.0
python-dotenv>=1.0.0
playwright-stealth>=1.0.5

# Зависимости для парсинга
aiohttp>=3.9.1
beautifulsoup4>=4.12.0

# Вспомогательные библиотеки
pydantic_settings>=2.0.0

# Добавленные зависимости
asyncio==3.4.3
aiosignal==1.3.1
python-telegram-bot==13.15
Pillow==9.5.0
aiofiles==23.1.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
matplotlib==3.7.2
prometheus-client==0.21.1
